requirements.txt
whisperx
fpdf2==2.8.3
python-dotenv==1.0.0
requests==2.31.0
